more changes to remove Io
