+++ fs/djs/parser/module.f.ts
@@ -9,12 +9,10 @@ import type { Fold } from '../../types/function/operator/module.f.ts'
 import type { DjsToken, DjsTokenWithMetadata } from '../tokenizer/module.f.ts'
 import { setReplace, at, type OrderedMap } from '../../types/ordered_map/module.f.ts'
 import { fromMap } from '../../types/object/module.f.ts'
-import type { Fs } from '../../io/module.f.ts'
 import type { AstArray, AstConst, AstModule, AstModuleRef } from '../ast/module.f.ts'
 import type { TokenMetadata } from '../../js/tokenizer/module.f.ts'
 
 export type ParseContext = {
-    readonly fs: Fs
     readonly complete: OrderedMap<Result<AstModule, string>>
     readonly stack: List<string>
 }
