+++ fs/text/sgr/module.f.ts
@@ -8,7 +8,6 @@
 // C0 control codes
 // https://en.wikipedia.org/wiki/ANSI_escape_code#C0_control_codes
 
-import type { Io, Writable } from "../../io/module.f.ts"
 import { write, type Write, type WriteConsoles, type NodeProgramOptions } from '../../effects/node/module.f.ts'
 import { type Effect } from '../../effects/module.f.ts'
 import { utf8 } from "../module.f.ts"
@@ -82,26 +81,6 @@ export type CsiConsole = (s: string) => void
 
 const str = (isTTY: boolean) => (s: string) =>
     isTTY ? s : s.replace(/\x1b\[[0-9;]*m/g, '')
-
-/**
- * Creates a TTY-aware console function. Appends `\n` to each string before writing.
- *
- * For TTY destinations, ANSI SGR sequences are preserved.
- * For non-TTY destinations, ANSI SGR sequences are stripped.
- *
- * @param io - Runtime IO bindings.
- * @returns A function that targets a writable stream.
- */
-export const console = ({ fs: { writeSync } }: Io) => (w: Writable): CsiConsole => {
-    const toStr = str(w.isTTY)
-    return (s: string) => writeSync(w.fd, toStr(s) + '\n')
-}
-
-/** Writes to process stdout using a TTY-aware CSI console. */
-export const stdio = (io: Io): CsiConsole => console(io)(io.process.stdout)
-
-/** Writes to process stderr using a TTY-aware CSI console. */
-export const stderr = (io: Io): CsiConsole => console(io)(io.process.stderr)
 
 /**
  * Effect-based TTY-aware write. Strips ANSI SGR sequences when the target
